requirements.txt
alembic==1.13.1
aniso8601==9.0.1
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.9.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
flask-restplus==0.13.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
protobuf==5.26.1
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pymysql-pool==0.4.6
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
